Switched to using dill for serializing artifacts.
Able to save classes along with needing to import the class definition.

Use case: I make a custom class in a Jupyter notebook, and save it as an artifact.
Dill makes it possible to use that artifact in another notebook even though the
original definition is inaccessible inside another notebook